requirements.txt
ezSASRec
pandas
pandarallel